:recycle:refactoring redis viewcount update
