+++ src/main/java/com/project/cheerha/common/dto/ViewCountResponseDto.java
@@ -1,0 +1,7 @@
+package com.project.cheerha.common.dto;
+
+public record ViewCountResponseDto(
+   Long key,
+   Long Value
+) {
+}
