+++ src/main/java/com/project/cheerha/domain/jobopening/repository/JobOpeningRepository.java
@@ -11,10 +11,9 @@ import java.util.List;
 public interface JobOpeningRepository extends JpaRepository<JobOpening, Long>, JobOpeningRepositoryQuery {
 
     /**
-     * 채용공고 테이블에 있는 viewCount 컬럼에 집계테이블 viewcount 컬럼 값 만큼 더해서 값을
-     * 업데이트 하는 메서드입니다.
-     * @param id 집계테이블의 특정 채용공고 id
-     * @param viewCount  집계테이블의 특정 viewCount
+     * 레디스의 조회 수를 채용공고 RDB에 업데이트 할 때 사용하는 메서드입니다.
+     * @param id 레디스의 특정 채용공고 id (key)
+     * @param viewCount 레디스의 특정 viewCount (value)
      */
     @Modifying
     @Query("UPDATE JobOpening b SET b.viewCount = b.viewCount + :viewCount WHERE b.id = :id")
