+++ src/main/java/com/project/cheerha/domain/jobopening/schdeuler/JobOpeningViewCountTaskHandler.java
@@ -1,23 +1,23 @@
-package com.project.cheerha.domain.viewcount.scheduler;
+package com.project.cheerha.domain.jobopening.schdeuler;
 
+import com.project.cheerha.common.dto.ViewCountResponseDto;
+import com.project.cheerha.common.redis.viewcount.RedisViewCountManager;
 import com.project.cheerha.common.scheduler.TaskHandler;
 import com.project.cheerha.domain.jobopening.repository.JobOpeningRepository;
-import com.project.cheerha.domain.viewcount.repository.JobOpeningViewCountRepository;
+import java.util.List;
+import java.util.Map;
 import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
 import org.springframework.stereotype.Component;
 import org.springframework.transaction.annotation.Transactional;
-
-import java.util.List;
-import java.util.Map;
 
 @Slf4j
 @Component
 @RequiredArgsConstructor
 public class JobOpeningViewCountTaskHandler implements TaskHandler {
 
-    private final JobOpeningViewCountRepository jobOpeningViewCountRepository;
     private final JobOpeningRepository jobOpeningRepository;
+    private final RedisViewCountManager redisViewCountManager;
 
     @Override
     public String getTaskType() {
@@ -27,15 +27,13 @@ public class JobOpeningViewCountTaskHandler implements TaskHandler {
     @Override
     @Transactional
     public void handle(Map<String, Object> payload) {
-        List<Long> jobOpeningIdList = jobOpeningViewCountRepository.findDistinctViewedJobOpeningIdList();
-        for (Long jobOpeningId : jobOpeningIdList) {
-            Long viewCount = jobOpeningViewCountRepository.getViewCountByJobOpeningId(jobOpeningId);
-            log.info("🔍 JobOpening ID: {}, ViewCount: {}", jobOpeningId, viewCount);
+        List<ViewCountResponseDto> redisJobOpeningViewCountList = redisViewCountManager.findAllViewCount();
+        for (ViewCountResponseDto jobOpening : redisJobOpeningViewCountList) {
+            Long jobOpeningId = jobOpening.key();
+            Long viewCount = jobOpening.Value();
             if (viewCount != null && viewCount > 0) {
                 jobOpeningRepository.updateViewCount(jobOpeningId, viewCount);
-                log.info("✅ JobOpening ID {}: ViewCount {} 적용 완료", jobOpeningId, viewCount);
-                jobOpeningViewCountRepository.resetViewCountByJobOpeningId(jobOpeningId);
-                log.info("🔄 JobOpeningViewCount 초기화 완료 - ID: {}", jobOpeningId);
+                redisViewCountManager.resetViewCount(jobOpeningId);
             }
         }
         log.info("조회수 동기화 완료");
@@ -43,6 +41,7 @@ public class JobOpeningViewCountTaskHandler implements TaskHandler {
 
     @Override
     public long getScheduleIntervalMillis() {
-        return 3600000L; // 1시간
+       // return 3600000L; // 1시간
+        return 300000L; //5분
     }
 }
